+++ lib/gds_api/test_helpers/account_api.rb
@@ -69,6 +69,19 @@ module GdsApi
         )
       end
 
+      def stub_update_user_by_subject_identifier(subject_identifier:, email: nil, email_verified: nil, old_email: nil, old_email_verified: nil)
+        stub_account_api_request(
+          :patch,
+          "/api/oidc-users/#{subject_identifier}",
+          with: { body: hash_including({ email: email, email_verified: email_verified }.compact) },
+          response_body: {
+            sub: subject_identifier,
+            email: email || old_email,
+            email_verified: email_verified || old_email_verified,
+          },
+        )
+      end
+
       def stub_account_api_has_email_subscription(**options)
         stub_account_api_request(
           :get,